requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
accelerate=0.18.0=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
altair=4.2.2=pypi_0
anyio=3.6.2=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.01.10=hca03da5_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
contourpy=1.0.7=pypi_0
controlnet-aux=0.0.3=pypi_0
cycler=0.11.0=pypi_0
dataclasses-json=0.5.7=pypi_0
diffusers=0.16.1=pypi_0
dino=0.0.6=pypi_0
einops=0.6.1=pypi_0
entrypoints=0.4=pypi_0
fastapi=0.95.1=pypi_0
ffmpy=0.3.0=pypi_0
filelock=3.12.0=pypi_0
fonttools=4.39.3=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.4.0=pypi_0
ftfy=6.1.1=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gradio=3.24.1=pypi_0
gradio-client=0.1.4=pypi_0
grpcio=1.54.0=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.0=pypi_0
httpx=0.24.0=pypi_0
huggingface-hub=0.14.1=pypi_0
idna=3.4=pypi_0
imageio=2.28.1=pypi_0
importlib-metadata=6.6.0=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=pypi_0
langchain=0.0.101=pypi_0
lazy-loader=0.2=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.2=hca03da5_6
linkify-it-py=2.0.1=pypi_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
matplotlib=3.7.1=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h313beb8_0
networkx=3.1=pypi_0
numexpr=2.8.4=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.6=pypi_0
openapi-schema-pydantic=1.2.4=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h1a28f6b_0
orjson=3.8.11=pypi_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=py310hca03da5_0
protobuf=4.22.4=pypi_0
psutil=5.9.5=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydantic=1.10.7=pypi_0
pydub=0.25.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.10.0=hbdb9e5c_5
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h1a28f6b_0
regex=2023.3.23=pypi_0
requests=2.29.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scipy=1.10.1=pypi_0
segment-anything=1.0=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=66.0.0=py310hca03da5_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlalchemy=1.4.47=pypi_0
sqlite=3.41.2=h80987f9_0
starlette=0.26.1=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tifffile=2023.4.12=pypi_0
timm=0.6.13=pypi_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.13.3=pypi_0
toolz=0.12.0=pypi_0
torch=1.13.1=pypi_0
torchaudio=0.13.1=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.65.0=pypi_0
transformers=4.28.1=pypi_0
typing-extensions=4.5.0=pypi_0
typing-inspect=0.8.0=pypi_0
tzdata=2023.3=pypi_0
uc-micro-py=1.0.2=pypi_0
urllib3=1.26.15=pypi_0
uvicorn=0.22.0=pypi_0
wcwidth=0.2.6=pypi_0
websockets=11.0.2=pypi_0
werkzeug=2.3.3=pypi_0
wget=3.2=pypi_0
wheel=0.38.4=py310hca03da5_0
xz=5.2.10=h80987f9_1
yarl=1.9.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5a0b063_0
